requirements.txt
langchain
Flask
tensorflow
numpy
scikit-learn
joblib
streamlit
playwright
unstructured[docx]
python-dotenv
ipykernel
langchain-community
pypdf
bs4
langchain-text-splitters
langchain-openai
#chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
#langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo-search
pymupdf
arxiv
wikipedia
SQLAlchemy
validators==0.28.1
unstructured
pytube
numexpr
huggingface_hub
langchain_groq